# docs/swagger.yaml
definitions:
  bet.Bet:
    properties:
      id:
        type: string
      matches:
        items:
          $ref: '#/definitions/match.Match'
        type: array
      score:
        type: integer
    type: object
...
  match.Match:
    properties:
      date:
        type: string
      matchType:
        $ref: '#/definitions/match.MatchType'
      result:
        $ref: '#/definitions/match.MatchResult'
      round:
        type: integer
      stadium:
        type: string
      teamA:
        type: string
      teamB:
        type: string
    type: object
  match.MatchResult:
    properties:
      teamAScore:
        type: integer
      teamBScore:
        type: integer
    type: object
  match.MatchType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - MatchTypeGroup
    - MatchTypeKnockout
info:
  contact: {}
  title: Bolao 2026
  version: "1.0"
paths:
  /bets/{id}:
    get:
      parameters:
      - description: Bet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bet.Bet'
      tags:
      - Bet
...
swagger: "2.0"